requirements.txt
pytorch==1.4.0
transformers==2.8.0
configargparse==1.1
bidict==0.18.0
fire==0.2.1
spacy==2.0.12